modulo12/requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
openai==1.3.0 